requirements.txt
# Core dependencies
pytesseract==0.3.10
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.4

# Database
pandas==2.2.0

# File monitoring
watchdog==4.0.0

# Dashboard
streamlit==1.31.1
plotly==5.19.0

# Utilities
python-dateutil==2.8.2
